+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- changed: If broadcasting fails, return an error with a better summary.
+
 ## 4.36.2 (2025-02-12)
 
 - fixed: (Solana) Fix priority fee for undefined `networkFeeOption`